backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
anthropic>=0.40.0
chromadb>=0.5.0
praat-parselmouth>=0.4.4
soundfile>=0.12.0
numpy>=1.26.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
edge-tts>=6.0.0
pedalboard>=0.9.0
pyloudnorm>=0.1.0
scipy>=1.12.0
librosa>=0.10.0
